app/requirements.txt
fastapi
uvicorn[standard]
opencv-python
ultralytics
numpy
psycopg2-binary
sort-tracker
pydantic-settings